# api/openapi.yaml
openapi: 3.0.3
info:
  title: Live Streaming Tipping Platform API
  description: |
    Secure, scalable live-streaming platform where performers broadcast video, 
    viewers watch, and viewers tip virtual tokens in real time.
    
    ## Key Features
    - Real-time WebRTC streaming with SFU
    - Virtual token economy with instant tipping
    - Age verification and KYC compliance
    - Automated and human moderation
    - PCI-compliant payment processing
    
    ## Authentication
    All endpoints require JWT authentication except registration, login, and public endpoints.
    Include the JWT token in the Authorization header: `Bearer <token>`
    
    ## Rate Limiting
    - General API: 1000 requests/hour per user
    - Streaming endpoints: 100 requests/minute per user
    - Payment endpoints: 10 requests/minute per user
    - Tipping: 60 tips/minute per user
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@livepanty.com
  license:
    name: Proprietary
servers:
  - url: https://api.livepanty.com/v1
    description: Production server
  - url: https://staging-api.livepanty.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /streams:
    get:
      tags:
        - Streaming
      summary: Get live streams
      description: Retrieve list of currently live streams
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Live streams retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamList'

    post:
      tags:
        - Streaming
      summary: Create new stream
      description: Create a new streaming session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStreamRequest'
      responses:
        '201':
          description: Stream created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamSession'
  /streams/{streamId}:
    get:
      tags:
        - Streaming
      summary: Get stream details
      description: Retrieve detailed information about a specific stream
      parameters:
        - $ref: '#/components/parameters/StreamIdPath'
      responses:
        '200':
          description: Stream details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamDetails'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Streaming
      summary: Update stream
      description: Update stream information (title, description, etc.)
      parameters:
        - $ref: '#/components/parameters/StreamIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStreamRequest'
      responses:
        '200':
          description: Stream updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamDetails'

    delete:
      tags:
        - Streaming
      summary: End stream
      description: End the streaming session
      parameters:
        - $ref: '#/components/parameters/StreamIdPath'
      responses:
        '200':
          description: Stream ended successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Stream ended successfully"
  /streams/{streamId}/start:
    post:
      tags:
        - Streaming
      summary: Start streaming
      description: Mark stream as started and begin broadcasting
      parameters:
        - $ref: '#/components/parameters/StreamIdPath'
      responses:
        '200':
          description: Stream started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamSession'
...
  /moderation/report:
    post:
      tags:
        - Moderation
      summary: Report content or user
      description: Submit a report for inappropriate content or behavior
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '201':
          description: Report submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    type: string
                    format: uuid
                  message:
                    type: string
                    example: "Report submitted successfully"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    StreamIdPath:
      name: streamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Stream ID
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CreateStreamRequest:
      type: object
      required:
        - title
        - category
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "My Amazing Stream"
        description:
          type: string
          maxLength: 1000
        category:
          type: string
          example: "entertainment"
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        is_private:
          type: boolean
          default: false
        is_age_restricted:
          type: boolean
          default: true
        tip_enabled:
          type: boolean
          default: true
        chat_enabled:
          type: boolean
          default: true
        recording_enabled:
          type: boolean
          default: false
    StreamSession:
      type: object
      properties:
        stream_id:
          type: string
          format: uuid
        signaling_url:
          type: string
          format: uri
        sfu_room_id:
          type: string
        turn_servers:
          type: array
          items:
            $ref: '#/components/schemas/TURNServer'
        viewer_token:
          type: string
        expires_at:
          type: string
          format: date-time
    TURNServer:
      type: object
      properties:
        urls:
          type: array
          items:
            type: string
        username:
          type: string
        credential:
          type: string
    StreamDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
        host_id:
          type: string
          format: uuid
        host_name:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        is_private:
          type: boolean
        is_age_restricted:
          type: boolean
        tip_enabled:
          type: boolean
        chat_enabled:
          type: boolean
        status:
          type: string
          enum: [created, starting, live, ended, suspended, banned]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        duration_seconds:
          type: integer
        viewer_count:
          type: integer
        peak_viewer_count:
          type: integer
        total_tips_received:
          type: integer
        total_tokens_received:
          type: integer
        thumbnail_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
    StreamList:
      type: object
      properties:
        streams:
          type: array
          items:
            $ref: '#/components/schemas/StreamSummary'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    StreamSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        host_id:
          type: string
          format: uuid
        host_name:
          type: string
        title:
          type: string
        category:
          type: string
        status:
          type: string
        is_private:
          type: boolean
        viewer_count:
          type: integer
        total_tokens_received:
          type: integer
        thumbnail_url:
          type: string
          format: uri
        started_at:
          type: string
          format: date-time
    UpdateStreamRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        category:
          type: string
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        tip_enabled:
          type: boolean
        chat_enabled:
          type: boolean
...
    ReportRequest:
      type: object
      required:
        - target_type
        - target_id
        - report_type
        - reason
      properties:
        target_type:
          type: string
          enum: [user, stream, message, tip]
        target_id:
          type: string
          format: uuid
        report_type:
          type: string
          enum: [inappropriate_content, harassment, spam, underage, fraud, payment_issue, other]
        reason:
          type: string
          minLength: 10
          maxLength: 1000
        evidence_urls:
          type: array
          items:
            type: string
            format: uri
          maxItems: 5
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time